Drop noexcept from Box2D ops that delegate to IntervalD.
Box2D doesn't maintain the same guarantees w.r.t. infinite values,
so it's possible that accessing an x or y interval from a box will
throw.

Ideally we'll fix that some day, but these are not edge cases we hit
often (if ever), so it may be a low priority.