Add reflectedAbout methods to Box classes.
These are intended to be used as replacement implementations for the
flipLR and flipTB methods, but the behavior of those is different
for reasons I don't understand (they just seem wrong, but the
documentation is unclear); this is DM-21487.  While the behavior of
these methods makes sense (and is better documentated), it's not clear
we have a use case for them if they can't be used to back the flip
methods.