Add conversions between geom::Angle and sphgeom::Angle.
It's a shame that both exist, but I'm not up for the disruption of
fixing it.  The least we can do is provide conversions, as we already
do between geom::SpherePoint and sphgeom::LonLat.

The conversions in C++ are explicit because I didn't want to have to
think about the potential for surprising behavior when that interacts
with operator overloading.  That's not a problem in Python, so there
the conversion is implicit, and hence one can now pass geom::Angle to
pybind11-wrapped sphgeom routines that expect sphgeom::Angle.